Update
git-svn-id: https://svn.aros.org/svn/aros/trunk@3963 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ compiler/include/clib/mathieeesingbas_protos.h
@@ -19,57 +19,57 @@
 /*
     Prototypes
 */
-AROS_LP1(LONG, IEEESPAbs,
-    AROS_LPA(LONG, y, D0),
+AROS_LP1(float, IEEESPAbs,
+    AROS_LPA(float, y, D0),
     struct MathIeeeSingBasBase *, MathIeeeSingBasBase, 9, Mathieeesingbas)
 
-AROS_LP2(LONG, IEEESPAdd,
-    AROS_LPA(LONG, y, D0),
-    AROS_LPA(LONG, z, D1),
+AROS_LP2(float, IEEESPAdd,
+    AROS_LPA(float, y, D0),
+    AROS_LPA(float, z, D1),
     struct MathIeeeSingBasBase *, MathIeeeSingBasBase, 11, Mathieeesingbas)
 
-AROS_LP1(LONG, IEEESPCeil,
-    AROS_LPA(LONG, y, D0),
+AROS_LP1(float, IEEESPCeil,
+    AROS_LPA(float, y, D0),
     struct MathIeeeSingBasBase *, MathIeeeSingBasBase, 16, Mathieeesingbas)
 
 AROS_LP2(LONG, IEEESPCmp,
-    AROS_LPA(LONG, y, D0),
-    AROS_LPA(LONG, z, D1),
+    AROS_LPA(float, y, D0),
+    AROS_LPA(float, z, D1),
     struct MathIeeeSingBasBase *, MathIeeeSingBasBase, 7, Mathieeesingbas)
 
-AROS_LP2(LONG, IEEESPDiv,
-    AROS_LPA(LONG, y, D0),
-    AROS_LPA(LONG, z, D1),
+AROS_LP2(float, IEEESPDiv,
+    AROS_LPA(float, y, D0),
+    AROS_LPA(float, z, D1),
     struct MathIeeeSingBasBase *, MathIeeeSingBasBase, 14, Mathieeesingbas)
 
 AROS_LP1(LONG, IEEESPFix,
-    AROS_LPA(LONG, y, D0),
+    AROS_LPA(float, y, D0),
     struct MathIeeeSingBasBase *, MathIeeeSingBasBase, 5, Mathieeesingbas)
 
-AROS_LP1(LONG, IEEESPFloor,
-    AROS_LPA(LONG, y, D0),
+AROS_LP1(float, IEEESPFloor,
+    AROS_LPA(float, y, D0),
     struct MathIeeeSingBasBase *, MathIeeeSingBasBase, 15, Mathieeesingbas)
 
-AROS_LP1(LONG, IEEESPFlt,
+AROS_LP1(float, IEEESPFlt,
     AROS_LPA(LONG, y, D0),
     struct MathIeeeSingBasBase *, MathIeeeSingBasBase, 6, Mathieeesingbas)
 
-AROS_LP2(LONG, IEEESPMul,
-    AROS_LPA(LONG, y, D0),
-    AROS_LPA(LONG, z, D1),
+AROS_LP2(float, IEEESPMul,
+    AROS_LPA(float, y, D0),
+    AROS_LPA(float, z, D1),
     struct MathIeeeSingBasBase *, MathIeeeSingBasBase, 13, Mathieeesingbas)
 
-AROS_LP1(LONG, IEEESPNeg,
-    AROS_LPA(LONG, y, D0),
+AROS_LP1(float, IEEESPNeg,
+    AROS_LPA(float, y, D0),
     struct MathIeeeSingBasBase *, MathIeeeSingBasBase, 10, Mathieeesingbas)
 
-AROS_LP2(LONG, IEEESPSub,
-    AROS_LPA(LONG, y, D0),
-    AROS_LPA(LONG, z, D1),
+AROS_LP2(float, IEEESPSub,
+    AROS_LPA(float, y, D0),
+    AROS_LPA(float, z, D1),
     struct MathIeeeSingBasBase *, MathIeeeSingBasBase, 12, Mathieeesingbas)
 
 AROS_LP1(LONG, IEEESPTst,
-    AROS_LPA(LONG, y, D0),
+    AROS_LPA(float, y, D0),
     struct MathIeeeSingBasBase *, MathIeeeSingBasBase, 8, Mathieeesingbas)
