+++ compiler/include/clib/mathieeesingtrans_protos.h
@@ -19,65 +19,65 @@
 /*
     Prototypes
 */
-AROS_LP1(LONG, IEEESPAcos,
-    AROS_LPA(LONG, y, D0),
+AROS_LP1(float, IEEESPAcos,
+    AROS_LPA(float, y, D0),
     struct Library *, MathIeeeSingTransBase, 20, Mathieeesingtrans)
 
-AROS_LP1(LONG, IEEESPAsin,
-    AROS_LPA(LONG, y, D0),
+AROS_LP1(float, IEEESPAsin,
+    AROS_LPA(float, y, D0),
     struct Library *, MathIeeeSingTransBase, 19, Mathieeesingtrans)
 
-AROS_LP1(LONG, IEEESPCos,
-    AROS_LPA(LONG, y, D0),
+AROS_LP1(float, IEEESPCos,
+    AROS_LPA(float, y, D0),
     struct Library *, MathIeeeSingTransBase, 7, Mathieeesingtrans)
 
-AROS_LP1(LONG, IEEESPCosh,
-    AROS_LPA(LONG, y, D0),
+AROS_LP1(float, IEEESPCosh,
+    AROS_LPA(float, y, D0),
     struct Library *, MathIeeeSingTransBase, 11, Mathieeesingtrans)
 
-AROS_LP1(LONG, IEEESPExp,
-    AROS_LPA(LONG, y, D0),
+AROS_LP1(float, IEEESPExp,
+    AROS_LPA(float, y, D0),
     struct Library *, MathIeeeSingTransBase, 13, Mathieeesingtrans)
 
-AROS_LP1(LONG, IEEESPFieee,
-    AROS_LPA(LONG, y, D0),
+AROS_LP1(float, IEEESPFieee,
+    AROS_LPA(float, y, D0),
     struct Library *, MathIeeeSingTransBase, 18, Mathieeesingtrans)
 
-AROS_LP1(LONG, IEEESPLog,
-    AROS_LPA(LONG, y, D0),
+AROS_LP1(float, IEEESPLog,
+    AROS_LPA(float, y, D0),
     struct Library *, MathIeeeSingTransBase, 14, Mathieeesingtrans)
 
-AROS_LP1(LONG, IEEESPLog10,
-    AROS_LPA(LONG, y, D0),
+AROS_LP1(float, IEEESPLog10,
+    AROS_LPA(float, y, D0),
     struct Library *, MathIeeeSingTransBase, 21, Mathieeesingtrans)
 
-AROS_LP2(LONG, IEEESPPow,
-    AROS_LPA(LONG, x, D1),
-    AROS_LPA(LONG, y, D0),
+AROS_LP2(float, IEEESPPow,
+    AROS_LPA(float, x, D1),
+    AROS_LPA(float, y, D0),
     struct Library *, MathIeeeSingTransBase, 15, Mathieeesingtrans)
 
-AROS_LP1(LONG, IEEESPSin,
-    AROS_LPA(LONG, y, D0),
+AROS_LP1(float, IEEESPSin,
+    AROS_LPA(float, y, D0),
     struct Library *, MathIeeeSingTransBase, 6, Mathieeesingtrans)
 
-AROS_LP1(LONG, IEEESPSinh,
-    AROS_LPA(LONG, y , D0),
+AROS_LP1(float, IEEESPSinh,
+    AROS_LPA(float, y , D0),
     struct Library *, MathIeeeSingTransBase, 10, Mathieeesingtrans)
 
-AROS_LP1(LONG, IEEESPSqrt,
-    AROS_LPA(LONG, y, D0),
+AROS_LP1(float, IEEESPSqrt,
+    AROS_LPA(float, y, D0),
     struct Library *, MathIeeeSingTransBase, 16, Mathieeesingtrans)
 
-AROS_LP1(LONG, IEEESPTan,
-    AROS_LPA(LONG, y, D0),
+AROS_LP1(float, IEEESPTan,
+    AROS_LPA(float, y, D0),
     struct Library *, MathIeeeSingTransBase, 8, Mathieeesingtrans)
 
-AROS_LP1(LONG, IEEESPTanh,
-    AROS_LPA(LONG, y, D0),
+AROS_LP1(float, IEEESPTanh,
+    AROS_LPA(float, y, D0),
     struct Library *, MathIeeeSingTransBase, 12, Mathieeesingtrans)
 
-AROS_LP1(LONG, IEEESPTieee,
-    AROS_LPA(LONG, y, D0),
+AROS_LP1(float, IEEESPTieee,
+    AROS_LPA(float, y, D0),
     struct Library *, MathIeeeSingTransBase, 17, Mathieeesingtrans)
 
 #endif /* CLIB_MATHIEEESINGTRANS_PROTOS_H */
